+++ routes/core/streams/create.int.test.js
@@ -44,4 +44,61 @@ describe('POST /streams', () => {
     expect(response.statusCode).toBe(400)
     expect(console.warn).toHaveBeenCalled()
   })
+
+  test('returns 201 when user creates a stream in a project he has a Member role', async () => {
+    const project = { id: 'foo', name: 'not my project', createdById: seedValues.otherUserId }
+    await models.Project.create(project)
+    await models.UserProjectRole.create({ user_id: seedValues.primaryUserId, project_id: project.id, role_id: seedValues.roleMember })
+
+    const requestBody = {
+      name: 'Trail 40',
+      latitude: 10.123,
+      longitude: 101.456,
+      project_id: 'foo'
+    }
+
+    const response = await request(app).post('/').send(requestBody)
+
+    expect(response.statusCode).toBe(201)
+    const id = response.header.location.replace('/streams/', '')
+    const stream = await models.Stream.findByPk(id)
+    expect(stream.name).toBe(requestBody.name)
+    expect(stream.createdById).toBe(seedValues.primaryUserId)
+  })
+  test('returns 201 when user creates a stream in a project he has an Admin role', async () => {
+    const project = { id: 'foo', name: 'not my project', createdById: seedValues.otherUserId }
+    await models.Project.create(project)
+    await models.UserProjectRole.create({ user_id: seedValues.primaryUserId, project_id: project.id, role_id: seedValues.roleAdmin })
+
+    const requestBody = {
+      name: 'Trail 41',
+      latitude: 10.123,
+      longitude: 101.456,
+      project_id: 'foo'
+    }
+
+    const response = await request(app).post('/').send(requestBody)
+
+    expect(response.statusCode).toBe(201)
+    const id = response.header.location.replace('/streams/', '')
+    const stream = await models.Stream.findByPk(id)
+    expect(stream.name).toBe(requestBody.name)
+    expect(stream.createdById).toBe(seedValues.primaryUserId)
+  })
+  test('returns 403 when user creates a stream in a project he has only a Guest role', async () => {
+    const project = { id: 'foo', name: 'not my project', createdById: seedValues.otherUserId }
+    await models.Project.create(project)
+    await models.UserProjectRole.create({ user_id: seedValues.primaryUserId, project_id: project.id, role_id: seedValues.roleGuest })
+
+    const requestBody = {
+      name: 'Trail 42',
+      latitude: 10.123,
+      longitude: 101.456,
+      project_id: 'foo'
+    }
+
+    const response = await request(app).post('/').send(requestBody)
+
+    expect(response.statusCode).toBe(403)
+  })
 })